requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
Brotli==1.1.0
bs4==0.0.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cssselect2==0.7.0
Flask==3.0.3
Flask-Script==2.0.6
fonttools==4.53.0
html5lib==1.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pillow==10.3.0
pycparser==2.22
pydyf==0.10.0
pyphen==0.15.0
requests==2.32.3
six==1.16.0
soupsieve==2.5
tinycss2==1.3.0
urllib3==2.2.2
weasyprint==62.2
webencodings==0.5.1
Werkzeug==3.0.3
zopfli==0.2.3
